+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 ## 1.0.0-beta.7
 ### Breaking
 * remove second argument from `filter` callback
+* change mapping `&` from `make-object` to `object`
+* parser is now async
 ### Features
-* add handling of quasiquote vector `` `#(1 2 ,(+ 1 2)) `` using syntax extensions
 * new `random` procedure
 ### Bugfix
 * fix write vs display of character atoms
 * fix mixed case of named characters (e.g.: `#\Space`)
 * fix `do` macro scope according to R7RS
+* fix vector literals inside quoted list [#79](https://github.com/jcubic/lips/issues/79)
 
 ## 1.0.0-beta.6
 ### Breaking